get dispatcher to (almost) compile. z3 version is unknown and does not link for me...

+++ libalf/include/libalf/algorithm_RPNI.h
@@ -233,7 +233,7 @@ class RPNI : public learning_algorithm<answer> {
 		}}}
 
 		virtual void generate_statistics(void) {
-			this->statistics["memory.bytes"] = sizeof(this);
+			this->statistics["memory.bytes"] = (int)sizeof(this);
 		}
 
 		virtual bool sync_to_knowledgebase()
